add integration test for merge container nodes
