+++ integration-test/expected-tree/merge-node/level1-container-node1/level2-container-node2/fileD.md
@@ -1,0 +1,4 @@
+---
+title: FileD
+---
+# Content
