+++ integration-test/expected-tree/merge-node/level1-container-node2/file1.md
@@ -1,0 +1,4 @@
+---
+title: File1
+---
+# Content
