+++ integration-test/expected-tree/merge-node/level1-container-node2/fileY.md
@@ -1,0 +1,4 @@
+---
+title: FileY
+---
+# Content
