+++ integration-test/manifest
@@ -33,6 +33,30 @@ structure:
       nodes:
       - nodesSelector:
           path: https://github.com/{{$org}}/docforge/blob/{{$version}}/integration-test/tested-doc/html-tests/testedDir
+- name: merge-node
+  nodes:
+  - name: level1-container-node1
+    nodes:
+    - name: level2-container-node1
+      nodes:
+      - name: level3-container-node1
+        nodes:
+        - name: file1
+          source: https://github.com/{{$org}}/docforge/blob/{{$version}}/integration-test/tested-doc/merge-test/testFile.md
+        - name: file2
+          source: https://github.com/{{$org}}/docforge/blob/{{$version}}/integration-test/tested-doc/merge-test/testFile.md
+    - name: level2-container-node2
+      nodes:
+      - name: file1
+        source: https://github.com/{{$org}}/docforge/blob/{{$version}}/integration-test/tested-doc/merge-test/testFile.md
+  - name: level1-container-node2
+    nodes:
+    - name: file1
+      source: https://github.com/{{$org}}/docforge/blob/{{$version}}/integration-test/tested-doc/merge-test/testFile.md
+  nodesSelector:
+    path: https://github.com/{{$org}}/docforge/blob/{{$version}}/integration-test/tested-doc/merge-test/testSelector.yaml
+nodesSelector:
+  path: https://github.com/{{$org}}/docforge/blob/{{$version}}/integration-test/tested-doc/merge-test/testDir
 links:
     downloads:
         scope:
