+++ integration-test/tested-doc/merge-test/testFile.md
@@ -1,0 +1,1 @@
+# Content
