+++ integration-test/tested-doc/merge-test/testSelector.yaml
@@ -1,0 +1,21 @@
+{{ $org := .org}}
+{{ $version := .version}}
+structure:
+- name: level1-container-node1
+  nodes:
+  - name: level2-container-node1
+    nodes:
+    - name: level3-container-node1
+      nodes:
+      - name: fileY1
+        source: https://github.com/{{$org}}/docforge/blob/{{$version}}/integration-test/tested-doc/merge-test/testFile.md
+      - name: fileY2
+        source: https://github.com/{{$org}}/docforge/blob/{{$version}}/integration-test/tested-doc/merge-test/testFile.md
+  - name: level2-container-node2
+    nodes:
+    - name: fileY
+      source: https://github.com/{{$org}}/docforge/blob/{{$version}}/integration-test/tested-doc/merge-test/testFile.md
+- name: level1-container-node2
+  nodes:
+  - name: fileY
+    source: https://github.com/{{$org}}/docforge/blob/{{$version}}/integration-test/tested-doc/merge-test/testFile.md
