+++ pkg/api/parser.go
@@ -5,13 +5,28 @@
 package api
 
 import (
+	"bytes"
 	"gopkg.in/yaml.v3"
+	"html/template"
+	"strings"
 )
+
+// flagsVars variables for template resolving
+var flagsVars map[string]string
+
+// SetFlagsVariables initialize flags variables
+func SetFlagsVariables(vars map[string]string) {
+	flagsVars = vars
+}
 
 // Parse is ...
 func Parse(b []byte) (*Documentation, error) {
+	blob, err := resolveVariables(b, flagsVars)
+	if err != nil {
+		return nil, err
+	}
 	var docs = &Documentation{}
-	if err := yaml.Unmarshal(b, docs); err != nil {
+	if err = yaml.Unmarshal(blob, docs); err != nil {
 		return nil, err
 	}
 	return docs, nil
@@ -27,4 +42,21 @@ func Serialize(docs *Documentation) (string, error) {
 		return "", err
 	}
 	return string(b), nil
+}
+
+func resolveVariables(manifestContent []byte, vars map[string]string) ([]byte, error) {
+	var (
+		tmpl *template.Template
+		err  error
+		b    bytes.Buffer
+	)
+	tplFuncMap := make(template.FuncMap)
+	tplFuncMap["Split"] = strings.Split
+	if tmpl, err = template.New("").Funcs(tplFuncMap).Parse(string(manifestContent)); err != nil {
+		return nil, err
+	}
+	if err = tmpl.Execute(&b, vars); err != nil {
+		return nil, err
+	}
+	return b.Bytes(), nil
 }
